Full function signature in rST (functions & methods)
The only part we need to strip from the signature is the 'function' one,
sphinxcontrib-phpdomains handles pretty well the rest of the signature.
This adds a lot more details to the docs, by showing where the
parameters are, their default values and the optional return value Type
hint.

This also allows to visually differentiate instance methods from class
methods.